+++ YACReaderLibrary/ip_config_helper.h
@@ -1,0 +1,4 @@
+#ifndef YR_IP_CONFIG_HELPER
+#define YR_IP_CONFIG_HELPER
+QList<QString> getIpAddresses();
+#endif