+++ CHANGELOG.md
@@ -7,7 +7,8 @@ Version counting is based on semantic versioning (Major.Feature.Patch)
 ### YACReaderLibrary
 * Fix scroll in grid views when using Qt6 builds.
 * Fix deleting metadata from comics also deleted the number of pages info.
-* Use https://github.com/nayuki/QR-Code-generator instead of libqrencode for QR code generation
+* Use https://github.com/nayuki/QR-Code-generator instead of libqrencode for QR code generation.
+* Do not accept empty values for the server port in the server settings dialog.
 
 ### Server
 * New search API that exposes the search engine.