+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Version counting is based on semantic versioning (Major.Feature.Patch)
 
 ## WIP
 
+## 9.12
+
 ### YACReaderLibrary
 * Fix scroll in grid views when using Qt6 builds.
 * Fix deleting metadata from comics also deleted the number of pages info.